extract `parseHttpRequest` into separate class
